Use |\zf@fontname| in messages.
Use |\zf@fontname| instead of |\fontname\zf@basefont| in fontspec
messages. This is a mere cosmotic change; luaotfload stores the full
pathname in the font's name field which leads to a bit ugly messages.
If there is reason against using |\zf@fontname| with xelatex, we can
make this luatex specific.