\@@_declare_shapes_slanted:nn setup a `scsl -> scit` substitution
If a slanted font is not provided, fontspec setup automatic sl -> it
substitutions in \@@_declare_shape_slanted:nn. But if a small caps
font was provided, an scit shape is defined but there is no
corresponding scsl -> scit substitution.

Fix this by adapting the code of \@@_declare_shapes_smcaps:nn into
\@@_declare_shape_slanted:nn.

Use a similar fix in \@@_declare_shapes_bx:nn for the bx/scsl -> bx/scit
substitution.