\__fontspec_check_script:Nn fails in some cases
If for some reason \curr@fontshape is not the really
the shape of the currently selected font, e.g. due
to a missing shape, and LuaTeX is the used engine,
the expansion of '\curr@fontshape/\f@size' is not
the name of the font switch that needs to be checked.
Example:

\documentclass{article}
\usepackage{fontspec}
\setmainfont{David CLM}
\begin{document}
\scshape % comment this line to see the effect
\ExplSyntaxOn
\fontspec_if_script:nTF{hebr}{yes}{no}
\end{document}

Note how this example returns true with LuaTeX but
false with XeTeX.

It would be more reliable to use \the\font to get
the name of the switch of the currently selected font.